LG-9164: Centralize address search hooks/types/utils (#9009)
* changelog: Internal, In-Person Proofing, Centralized utils, types, and hooks

* Properly import type definitions
