+++ app/javascript/packages/document-capture/components/in-person-location-full-address-entry-post-office-search-step.tsx
@@ -3,7 +3,8 @@ import { useI18n } from '@18f/identity-react-i18n';
 import { Alert, PageHeading } from '@18f/identity-components';
 import { request } from '@18f/identity-request';
 import { forceRedirect } from '@18f/identity-url';
-import { transformKeys, snakeCase, LocationQuery } from '@18f/identity-address-search';
+import { transformKeys, snakeCase } from '@18f/identity-address-search';
+import type { LocationQuery } from '@18f/identity-address-search/types';
 import FullAddressSearch from './in-person-full-address-search';
 import BackButton from './back-button';
 import AnalyticsContext from '../context/analytics';
